+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_container_client.hpp
@@ -211,7 +211,7 @@ namespace Azure { namespace Storage { namespace Blobs {
      * enumerating the blobs segment by segment. Blobs are ordered lexicographically by name.
      *
      * @param options Optional parameters to execute this function.
-     * @return A ListBlobsFlatSegmentResult describing a segment of the blobs in the container.
+     * @return A ListBlobsSinglePageResult describing a segment of the blobs in the container.
      */
     Azure::Core::Response<Models::ListBlobsSinglePageResult> ListBlobsSinglePage(
         const ListBlobsSinglePageOptions& options = ListBlobsSinglePageOptions()) const;
@@ -227,7 +227,7 @@ namespace Azure { namespace Storage { namespace Blobs {
      * @param delimiter This can be used to to traverse a virtual hierarchy of blobs as though it
      * were a file system. The delimiter may be a single character or a string.
      * @param options Optional parameters to execute this function.
-     * @return A ListBlobsByHierarchySegmentResult describing a segment of the blobs in the
+     * @return A ListBlobsByHierarchySinglePageResult describing a segment of the blobs in the
      * container.
      */
     Azure::Core::Response<Models::ListBlobsByHierarchySinglePageResult>