+++ sdk/storage/azure-storage-blobs/CHANGELOG.md
@@ -56,6 +56,7 @@
 - `ListBlobsFlatSegment` is renamed to `ListBlobsSinglePage`.
 - `ListBlobsByHierarchySegment` is renamed to `ListBlobsByHierarchySinglePage`.
 - `ListBlobContainersSegment` is renamed to `ListBlobContainersSinglePage`.
+- `FindBlobsByTags` is renamed to `FindBlobsByTagsSinglePage`.
 - `MaxResults` in list APIs are renamed to `PageSizeHint`.
 
 ## 12.0.0-beta.5 (2020-11-13)